Group focused class for holding information about sets of SNPs defined by their mutual LD

groups and tags are two structures for holding the same information, depending on whether your focus is on the sets of SNPs or their index members. It is easy to convert one to another and perhaps, in future, one class may be deprecated.

groups and tags are two structures for holding the same information, depending on whether your focus is on the sets of SNPs or their index members. It is easy to convert one to another and perhaps, in future, one class may be deprecated.

Slots

tags

character vector giving tag SNPs. Each tag indexes one group of SNPs

.Data

list of character vectors giving the SNP membership of each group

tags

character vector giving tag SNPs, one per SNP in in .Data, repeated as necessary

.Data

character vector giving SNPs included in this tags object

Examples

new("groups")

#> groups object, containing 0 SNPs in 0 groups.

new("tags")

#> tags object, containing 0 SNPs in 0 groups.