Summarise a snpmod object

guess.summ

guess.summ(results, groups = NULL, snps = NULL, snp.data = NULL,
  position = "position", tag.thr = 0.8, pp.thr = 0.01,
  method = "complete")

Arguments

results
an object of class snpmod, or a named list of such objects

groups
object of class groups. If supplied, all SNPs here will be summarised, and grouped according to this structure

snps
data.frame giving details of the SNPs

snp.data
if groups is missing, tag groups are determined from this object of class SnpMatrix

position
name of column in snps data.frame giving SNP position. default="position"

tag.thr
if groups is missing, threshold at which to tag

pp.thr
if groups is missing, threshold above which SNPs are selected for summary.

method
if groups is missing, method to determine tag groups using heirarchical clustering, default is "complete"

Value

data.frame